requirements.txt
torch>=1.13.0
hydra-core>=1.3.0
omegaconf>=2.1.0
accelerate>=1.0.0
tqdm>=4.0.0
